[MOD] Get file from firebase url
